Pending discussion about Phonon::Mrl as thing to take care of appropriate encoding and percent encoding here is nice workaround code for problems with special characters in file names such as '#'.
Basically the problem is that most media frameworks require perecent encoded URIs at the same time we need to ensure that local files actually get percent-encoded from their local encoded rather than UTF-8. The problem here however is that QUrl's toEncoded first converts everything to UTF-8 and then applies percent encoding on it, which of course breaks if the system encoding is ISOsomething. At the same time QByteArray's toPercentEncoding does not care about URLs and has a rather greeded encoding policy converting / and : and @ and other valid parts of URLs as well. So, both existing implementations render the URI broken in various use cases which suggests we should make one that suites our needs. Until we get an Mrl (which would at least be used by GStreamer and VLC) a local resolution for GStreamer.

BUG: #266727