Redo the name and description in the device info
Formerly, the name of the device was put as the gstId, and the device
access list wasn't used. Now the gstId can be obtained from the device
access list, and the name of the device can be the name shown to the
user (preferred name).

The description is somewhat similar to that from the KDE platform
plugin, and it gives an idea about the sound system and device id.