fix playback of discs on multi-drive setups
from the playbin docs:
The device to use for CDs/DVDs needs to be set on the source element
playbin creates before it is opened. The most generic way of doing
this is to connect to playbin's "source-setup" (or "notify::source")
signal, which will be emitted by playbin when it has created the source
element for a particular URI.

so we are now doing exactly that.

gstreamer being the great software that it is does not mandate using the
URI to set the device (i.e. use it as MRL like every other framework in
the world), so setting the property is the only reliable way to achieve
this.

note that I have no idea if there is a way to check whether a property is
present on a gobject so I am simply setting it whenever the source is a
disc. feel free to improve that.

also the callback handler needs some serious refactoring

CCMAIL: tdfischer@fedoraproject.org