do not unlock not locked mutexes, prevent crashing with qt5
handleEOS does not actually need mutex protection, or well, if it did then
unlocking would be counter productive.
from high level testing it seems that handleEOS only ought to be called
after handleAboutToFinish returned, so at that point we are done with
decision making. the other way is to have abouttonfinish waiting and
issue a state change which ought to trigger EOS while ATF is still pending
, in that case EOS will interrupt ATF which will result in essentially
the same result anyway.

eitherway, nothing to lock, so not unlocking is the way to go