Only call QApplication::syncX() from the main thread
This fixes a regression introduced in 2db4c4. In order to pass WID to gstreamer
synchronously, a Qt::DirectConnection between Pipeline signal and a VideoWidget
slot must be created. However the signal from Pipeline can be emitted from an
arbitrary thread, so we can't call QApplication::syncX() directly from the slot
callchain. This patch moves the call to VideoWidget::syncX() slot that is called
from X11Renderer via queued invocation to make sure it's always executed from the
main thread.

FIXED-IN: 4.7.3
BUG: 327367
BUG: 332977
CCBUG: 305333