Fix more memory leaks and optimize memory management
Calling gst_object_ref_sink() and gst_object_ref() means that we were holding two references to
the object, so obviously we were leaking the references later on, when we did only one unref.

Some structures have to be freed using special methods from Gst 1.0, so we use them to prevent
more reference or memory leaks.

Finally, this change refactors AbstractRenderer, Effect and MediaNode a little so that subclasses
must pass and access the Gst pointers (sinks, bins, ...) via setters and getters so that we can
propertly manage ownership of these objects and don't leak references or accidentally try to free
the object twice.