AudioDataOutput: flush m_pendingData when channels count changes
When channel count changes and m_pendingData is not empty, the new data
that we append to m_pendingData mess it up, because then we are no longer
able to correctly decode to data into m_channelBuffers.

This change makes the processBuffer() method to flush the pending data into
channel buffers and emit dataReady() signal when the channel count changes.
Only after that it continues to process the newly arrived data with the
new channel count).

BUG: 331916
FIXED-IN: 4.8